Update 'standard' to fix audit complaint
'standard' which is used for style linting uses the package 'ansi-regex'
which has a medium-level DDoS bug. There's no impact here, the package
is only run by a developer running `npm run lint`, and the input is the
source code, which is implicitly trusted if you're running that.

CVE-2021-3807 - ansi-regex is vulnerable to Inefficient Regular
                Expression Complexity